fix(safety,modules,docs): BUG-859..867 — nine holes, and the multi-site tax again
Audit sweep over the whole tree. Every row was reproduced on a pristine
`git archive HEAD` build BEFORE being fixed, every regression test was verified
to flip ACCEPT -> REJECT against that same build, and every over-rejection had
its corpus cost measured rather than argued. `make check` exit 0: 1345 .zer
tests, 139 module tests, all seven audit gates plus the new one, sink matrix
CLEAN.

Root-cause pattern, for the third wave running: eight of the nine are the SAME
semantic question answered independently at N sites with one site wrong, so two
spellings of one program disagree — `retain(&l)` rejected while `fp(&l)`
compiles, `free(buf[0..])` rejected while `free(s.buf[0..])` compiles, a local
funcptr rejected while its global sibling is not. Where a fix had a durable
form it is ONE query used at every site, not another arm.

Accept-unsafe:
- BUG-859 @cstr's destination gate was a stack of NEGATIVE tests, so an INTEGER
  destination matched none and was accepted by default: `@cstr(gi, sl)` emitted
  `memcpy(gi, ...)` and ran. That is the integer-to-pointer conversion the
  language says cannot be spelled without @inttoptr + mmio — a breach of the
  grammar-closure claim. Replaced with a positive gate.
- BUG-859 a negative CONSTANT into an unsigned destination (`u32 a = -1;`
  silently 4294967295) at all eight value-flow sinks. Unary minus is typed
  `result = operand`, so `-1` carries the literal's u32 and type_equals
  succeeds everywhere; nothing checked the value.
- BUG-861 @truncate / @saturate could FORGE an out-of-variant enum and the
  switch silently ran its LAST arm. `type_is_integer` answers TRUE for
  TYPE_ENUM, so both slipped their own "target must be an integer" gates.
  Redirected to @bitcast, which TRACKS (BUG-843's runtime variant guard).
- BUG-862 free() of a slice of an INLINE ARRAY through a struct field reached
  libc free() on a stack address. The gate asked about the ROOT SYMBOL's type;
  the durable question is about the SLICED OBJECT.
- BUG-863 the funcptr-call keep worst-case tested the BARE param kind, so `?*T`
  and a by-value struct carrying a pointer let a stack pointer reach a
  retaining callback. CONFIRMED BY ASAN as stack-use-after-return from pure
  safe ZER. Widened to a recursive carrier query.
- BUG-864 a global funcptr CALLED and never assigned jumped through NULL —
  silent on bare metal, where address 0 is usually mapped. The obvious fix
  (copy the local rule) was measured and BACKED OUT because it rejects the
  callback-registry idiom; the question moved into the whole-file pass.

Silent at compile time, loud at the backend:
- BUG-860 spawn had NO arity check; extras were dropped and only GCC objected.
- BUG-865 `spawn <imported function>` could not LINK — the wrapper spelled the
  target name RAW at five sites. Nothing in the suite had ever spawned an
  imported function.
- BUG-866 a container stamped across a module boundary was emitted once per
  module pass: "redefinition of 'struct Box_Item'".
- BUG-867 `const [*]u8 BANNER = "boot ok";` was undeclarable at file scope, and
  said so with the SAME type on both sides. Three copies of the qualifier fold
  collapse to one query.

Compiler-internal defects found in build warnings: lower_expr could fall off
the end and return an indeterminate local id; a `%.*s` fed a size_t; a calloc
count GCC's range analysis admits as negative.

Tooling:
- tools/audit_reference_examples.sh (new, in `make check`): every ```zer block
  in reference.md must parse as a file or as a function body. Verified
  non-vacuous by two injections.
- tools/ub_sweep.sh (new, standalone): builds one emitted .c four ways and
  compares answers — the method that pinned BUG-845.
- audit_walker_fields.sh decided membership on PROSE (it counted the function
  name in comments), so editing a comment moved a walker in or out of the
  audited set. Now requires an actual call in comment-stripped code; the two
  functions that leave the set are not recursive and were never in scope.

Docs: reference.md intrinsic coverage 41 -> 139, signatures MEASURED by probing
the compiler. 98 were undocumented, including the entire MMU, TLB, cache, port
I/O, control-register and privileged-transition families.

Repo hygiene: untrack 566 build artifacts under the test trees, identified by
CONTENT (558 ELF, 4 .exe, 4 files headed "Generated by ZER compiler"), and
replace the hand-maintained ignore name-list with a structural rule. `git
status` is usable for review again. rust_tests/qemu/link.ld and startup.c look
like artifacts and are genuine sources; both stay tracked.

+++ zercheck_ir.c
@@ -6274,7 +6274,10 @@ bool zercheck_ir(ZerCheck *zc, IRFunc *func) {
         func->ast_node->kind == NODE_FUNC_DECL) {
         fprintf(stderr, "ZCIR: building=%d fn='%.*s' pc=%d blocks=%d sumcount=%d\n",
             zc->building_summary,
-            func->ast_node->func_decl.name_len,
+            /* `%.*s` reads an `int` from varargs; `name_len` is a `size_t`, so
+             * this was a varargs type mismatch (formally UB, and wrong on any
+             * ABI where the two are passed differently). */
+            (int)func->ast_node->func_decl.name_len,
             func->ast_node->func_decl.name ? func->ast_node->func_decl.name : "?",
             func->ast_node->func_decl.param_count, func->block_count,
             zc->summary_count);
